modularize apollo providers
