+++ libs/driver/shopify/src/graphql/apollo-links/http-link.ts
@@ -1,0 +1,8 @@
+import { inject } from '@angular/core';
+import { ApolloLink } from '@apollo/client/core';
+import { HttpLink } from 'apollo-angular/http';
+
+export function createHttpLink(uri: string): ApolloLink {
+  const httpLink = inject(HttpLink);
+  return httpLink.create({ uri });
+}
