+++ libs/driver/shopify/src/graphql/provide-shopify-apollo-driver.ts
@@ -1,34 +1,22 @@
-import { inject } from '@angular/core';
 import { InMemoryCache } from '@apollo/client/cache';
-import { ApolloLink } from '@apollo/client/core';
-import { setContext } from '@apollo/client/link/context';
-import { onError } from '@apollo/client/link/error';
+import { from } from '@apollo/client/core';
 import { provideApollo } from 'apollo-angular';
-import { HttpLink } from 'apollo-angular/http';
+
+import {
+  createErrorLink,
+  createAuthLink,
+  createHttpLink,
+} from './apollo-links/public_api';
 
 export function provideShopifyApolloDriver(domain: string, accessToken: string) {
   const uri = `${domain}/api/2025-01/graphql.json`;
-  return provideApollo(() => {
-    const httpLink = inject(HttpLink);
-    const errorLink = onError(({ graphQLErrors, networkError }) => {
-      if (graphQLErrors) {
-        graphQLErrors.forEach(({ message, locations, path }) => {
-          console.error(`[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`);
-        });
-      }
-      if (networkError) {
-        console.error(`[Network error]: ${networkError}`);
-      }
-    });
-    const authLink = setContext(() => ({
-      headers: {
-        'Content-Type': 'application/json',
-        'X-Shopify-Storefront-Access-Token': accessToken,
-      },
-    }));
-    return {
-      link: ApolloLink.from([errorLink, authLink, httpLink.create({ uri })]),
-      cache: new InMemoryCache(),
-    };
-  });
+  return provideApollo(() => ({
+    link: from([
+      createErrorLink(),
+      createAuthLink(accessToken),
+      createHttpLink(uri),
+    ]),
+    cache: new InMemoryCache(),
+  }));
+
 }
